+++ src/jsMain/kotlin/at/posselt/pfrpg2e/kingdom/dialogs/AddEvent.kt
@@ -117,7 +117,7 @@ class AddEvent(
     width = 600,
     id = "kmEvents-${kingdomActor.uuid}",
     dataModel = AddEventsDataModel::class.js,
-    scrollable = arrayOf(".km-add-events"),
+    scrollable = setOf(".km-add-events"),
 ) {
     var search = ""