+++ src/jsMain/kotlin/com/foundryvtt/core/data/PolygonShapeData.kt
@@ -1,0 +1,11 @@
+package com.foundryvtt.core.data
+
+import com.foundryvtt.core.AnyObject
+import com.foundryvtt.core.abstract.DocumentConstructionContext
+
+external class PolygonShapeData(
+    data: AnyObject = definedExternally,
+    options: DocumentConstructionContext = definedExternally
+): BaseShapeData {
+    var points: Array<Int>
+}